requirements.txt
langgraph==0.2.48
langchain_anthropic==0.3.0
langchain_openai==0.2.8
langchain_community==0.3.7
python-dotenv==1.0.1
langgraph-checkpoint-sqlite==2.0.1
sounddevice
openai
scipy
elevenlabs